Fix size_t warnings when building for device

+++ apps/regression/graph/graph_view.cpp
@@ -32,7 +32,7 @@ void RegressionPlotPolicy::drawPlot(const Shared::AbstractPlotView *plotView,
     if (s == selectedSeries) {
       continue;
     }
-    int series = s;
+    size_t series = s;
     if (s == Store::k_numberOfSeries) {
       series = selectedSeries;
     }
